amazon-bedrock-image-generation-poc/requirements.txt
altair==5.1.2
attrs==23.1.0
awscli==1.29.60
blinker==1.6.2
boto3==1.28.60
botocore==1.31.60
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
docutils==0.16
gitdb==4.0.10
GitPython==3.1.41
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.2.0
pip==23.3
protobuf==4.24.4
pyarrow==14.0.1
pyasn1==0.5.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
rsa==4.7.2
s3transfer==0.7.0
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.18
validators==0.22.0
wheel==0.38.4
zipp==3.17.0
